+++ frontend/manage/src/hooks/api/project/useDeleteProject.ts
