+++ frontend/manage/src/hooks/api/user/useDeleteUser.ts
