+++ frontend/manage/src/hooks/index.ts
@@ -1,18 +1,18 @@
-export { useInput } from "./useInput";
-export { useUser } from "./useUser";
-export { useGetAllProjects } from "./useGetAllProjects";
-export { useGetProject } from "./useGetProject";
-export { useCreateProject } from "./useCreateProject";
-export { useCopyButton } from "./useCopyButton";
-export { useEditProject } from "./useEditProject";
-export { useEditUser } from "./useEditUser";
-export { useDeleteUser } from "./useDeleteUser";
-export { useGetCommentsOfProjectPerPage } from "./useGetCommentsOfProjectPerPage";
-export { useCommentList } from "./useCommentList";
-export { useCalendar } from "./useCalendar";
-export { useCommentPageIndex } from "./useCommentPageIndex";
-export { useDeleteComment } from "./useDeleteComment";
-export { useCommentStatisticsData } from "./useCommentStatisticsData";
-export { useDeleteProject } from "./useDeleteProject";
-export { useScrollFadeInOut } from "./useScrollFadeInOut";
-export { useToken } from "./useToken";
+export { useInput } from "./ui/useInput";
+export { useUser } from "./api/user/useUser";
+export { useGetAllProjects } from "./api/project/useGetAllProjects";
+export { useGetProject } from "./api/project/useGetProject";
+export { useCreateProject } from "./api/project/useCreateProject";
+export { useCopyButton } from "./ui/useCopyButton";
+export { useEditProject } from "./api/project/useEditProject";
+export { useEditUser } from "./api/user/useEditUser";
+export { useDeleteUser } from "./api/user/useDeleteUser";
+export { useGetCommentsOfProjectPerPage } from "./api/comment/useGetCommentsOfProjectPerPage";
+export { useCommentList } from "./api/comment/useCommentList";
+export { useCalendar } from "./ui/useCalendar";
+export { useCommentPageIndex } from "./ui/useCommentPageIndex";
+export { useDeleteComment } from "./api/comment/useDeleteComment";
+export { useCommentStatisticsData } from "./api/comment/useCommentStatisticsData";
+export { useDeleteProject } from "./api/project/useDeleteProject";
+export { useScrollFadeInOut } from "./ui/useScrollFadeInOut";
+export { useToken } from "./api/token/useToken";
