+++ frontend/manage/src/hooks/ui/useCopyButton.ts
