+++ frontend/manage/src/hooks/ui/useScrollFadeInOut.ts
