+++ frontend/reply-module/src/App.tsx
@@ -1,14 +1,15 @@
 import CommentArea from "./components/pages/CommentArea";
 import { BrowserRouter, Switch, Route, Redirect } from "react-router-dom";
 import OAuth from "./components/pages/OAuth";
+import { ROUTE } from "./constants/route";
 
 const App = () => {
   return (
     <BrowserRouter>
       <Switch>
-        <Route exact path="/" component={CommentArea} />
-        <Route exact path="/oauth/:provider" component={OAuth} />
-        <Redirect to="/" />
+        <Route exact path={ROUTE.HOME} component={CommentArea} />
+        <Route exact path={ROUTE.OAUTH} component={OAuth} />
+        <Redirect to={ROUTE.HOME} />
       </Switch>
     </BrowserRouter>
   );
